Time: 167 ms (69.18%) | Memory: 12.8 MB (37.84%) - LeetSync

+++ 204-count-primes/count-primes.cpp
@@ -8,7 +8,7 @@ public:
         isPrime[0] = isPrime[1] = false;
         for(int i = 2; i*i<n; i++){
             if(isPrime[i]){
-                for(long j = i * i; j< n; j+=i){
+                for(int j = i * i; j< n; j+=i){
                     isPrime[j] = false;
                 }
             }
